RobotRecieve manualed
