+++ Fedya/FANUC Client/FANUC Client.vcxproj.filters
@@ -72,6 +72,9 @@
     <ClCompile Include="RobotSend.cpp">
       <Filter>Файлы исходного кода\RobotConnection</Filter>
     </ClCompile>
+    <ClCompile Include="RobotRecieve.cpp">
+      <Filter>Файлы исходного кода\RobotConnection</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Field.hpp">
@@ -123,6 +126,9 @@
       <Filter>Заголовочные файлы\template difinitions\winSock</Filter>
     </ClInclude>
     <ClInclude Include="RobotSendDifinition.hpp">
+      <Filter>Заголовочные файлы\template difinitions\RobotConnection</Filter>
+    </ClInclude>
+    <ClInclude Include="RobotRecieveDifinition.hpp">
       <Filter>Заголовочные файлы\template difinitions\RobotConnection</Filter>
     </ClInclude>
   </ItemGroup>
