+++ Fedya/FANUC Client/RobotRecieve.cpp
@@ -1,0 +1,1 @@
+#include "RobotRecieve.hpp"
